[#1383] neofs-cli: move `object` command to a separate package
